// iki-client/types/movie.ts
export interface Movie {
    id: number;
    poster_link: string;
    series_title: string;
    released_year: string;
    certificate: string;
    runtime: string;
    genre: string;
    imdb_rating: number;
    overview: string;
    meta_score: number;
    director: string;
    star_one: string;
    star_two: string;
    star_three: string;
    star_four: string;
    number_of_votes: string;
    gross: string;
}